refactor(gateway, security): Use shared DTO instead local
Security service
- Removed RoleDto
- Minor changes in AuthServiceImpl

Gateway service
- Remove duplicated dependency for 'shared' service from pom
