+++ mp-security/src/main/java/ru/pavbatol/myplace/role/controller/AdminRoleController.java
@@ -7,8 +7,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import ru.pavbatol.myplace.role.dto.RoleDto;
 import ru.pavbatol.myplace.role.service.RoleService;
+import ru.pavbatol.myplace.shared.dto.security.role.RoleDto;
 
 import java.util.List;
