--- mp-security/src/main/java/ru/pavbatol/myplace/role/dto/RoleDto.java
@@ -1,10 +1,0 @@
-package ru.pavbatol.myplace.role.dto;
-
-import lombok.Value;
-import ru.pavbatol.myplace.role.model.RoleName;
-
-@Value
-public class RoleDto {
-    Long id;
-    RoleName roleName;
-}
