+++ mp-security/src/main/java/ru/pavbatol/myplace/role/mapper/RoleMapper.java
@@ -2,8 +2,8 @@ package ru.pavbatol.myplace.role.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
-import ru.pavbatol.myplace.role.dto.RoleDto;
 import ru.pavbatol.myplace.role.model.Role;
+import ru.pavbatol.myplace.shared.dto.security.role.RoleDto;
 
 import java.util.List;
