+++ mp-security/src/main/java/ru/pavbatol/myplace/user/dto/UserDtoResponse.java
@@ -1,7 +1,7 @@
 package ru.pavbatol.myplace.user.dto;
 
 import lombok.Value;
-import ru.pavbatol.myplace.role.dto.RoleDto;
+import ru.pavbatol.myplace.shared.dto.security.role.RoleDto;
 
 import java.util.HashSet;
 import java.util.Set;
